comfy/requirements.txt
accelerate==1.1.1
addict==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
boto3==1.35.72
botocore==1.35.72
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
clip-interrogator==0.6.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
deepdiff==8.0.1
diffusers==0.31.0
easydict==1.13
einops==0.8.0
eval_type_backport==0.2.0
evalidate==2.0.3
fal_client==0.5.6
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.3.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
insightface==0.7.3
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
kornia==0.7.4
kornia_rs==0.1.7
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
onnx==1.16.1
onnxruntime-gpu==1.19.2
open_clip_torch==2.29.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orderly-set==5.2.2
packaging==24.2
pandas==2.2.3
piexif==1.1.3
pillow==11.0.0
pip==24.3.1
platformdirs==4.3.6
pooch==1.8.2
prettytable==3.12.0
propcache==0.2.0
protobuf==5.29.0
psutil==6.1.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pygit2==1.16.0
PyMatting==1.1.13
pynvml==11.5.3
pyparsing==3.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
regex==2024.11.6
rembg==2.0.60
replicate==1.0.4
requests==2.32.3
rpds-py==0.21.0
s3transfer==0.10.4
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
segment-anything==1.0
sentencepiece==0.2.0
setuptools==75.6.0
simpleeval==1.0.3
simsimd==6.2.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
spandrel==0.4.0
stringzilla==3.11.0
surrealist==1.0.7
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.9
tokenizers==0.20.3
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchsde==0.2.6
torchvision==0.20.1+cu124
tqdm==4.67.0
trampoline==0.1.2
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.39
ultralytics-thop==2.0.12
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
yapf==0.43.0
yarl==1.17.2
zipp==3.21.0
